+++ src/__mocks/trading/GraphMock.ts
