+++ src/components/portfolio/PortfolioTable.tsx
@@ -12,7 +12,7 @@ import {DataTable} from "@/components/__common__/datatable/DataTable.tsx";
 import {DataTablePagination} from "@/components/__common__/datatable/DataTablePagination.tsx";
 import {generatePortfolioColumns} from "@/components/portfolio/PortfolioTableColumns.tsx";
 import {PortfolioData} from "@/types/exchange/portfolio-data.ts";
-import {portfolioDataMock} from "@/mocks/PortfolioDataMock.tsx";
+import {portfolioDataMock} from "@/__mocks/PortfolioDataMock.ts";
 
 import {
     Dialog,
