--- src/mocks/mokdatasExchange.tsx
@@ -1,157 +1,0 @@
-import { Exchange } from "@/types/exchange/exchange.ts";
-
-export const mokCountries: string[]= ["USA", "Indonesia", "Australia", "United States", "Ireland", "Italy", "India", "France", "Canada", "Poland", "Ukraine"];
-
-export const mokDataExchanges: Exchange[] = [
-    {
-      name: "Nasdaq",
-      acronym: "NASDAQ",
-      micCode: "XNAS",
-      country: "USA",
-      currency: "USD",
-      timeZone: "America/New_York",
-      openTime: "09:30",
-      closeTime: "16:00"
-    },
-    {
-      name: "Jakarta Futures Exchange (bursa Berjangka Jakarta)",
-      acronym: "BBJ",
-      micCode: "XBBJ",
-      country: "Indonesia",
-      currency: "Indonesian Rupiah",
-      timeZone: "Asia/Jakarta",
-      openTime: "09:00",
-      closeTime: "17:30"
-    },
-    {
-      name: "Asx - Trade24",
-      acronym: "SFE",
-      micCode: "XSFE",
-      country: "Australia",
-      currency: "Australian Dollar",
-      timeZone: "Australia/Melbourne",
-      openTime: "10:00",
-      closeTime: "16:00"
-    },
-    {
-      name: "Cboe Edga U.s. Equities Exchange Dark",
-      acronym: "EDGADARK",
-      micCode: "EDGD",
-      country: "United States",
-      currency: "United States Dollar",
-      timeZone: "America/New_York",
-      openTime: "09:30",
-      closeTime: "16:00"
-    },
-    {
-      name: "Clear Street",
-      acronym: "CLST",
-      micCode: "CLST",
-      country: "United States",
-      currency: "United States Dollar",
-      timeZone: "America/New_York",
-      openTime: "09:30",
-      closeTime: "16:00"
-    },
-    {
-      name: "Wall Street Access Nyc",
-      acronym: "WABR",
-      micCode: "WABR",
-      country: "United States",
-      currency: "United States Dollar",
-      timeZone: "America/New_York",
-      openTime: "09:30",
-      closeTime: "16:00"
-    },
-    {
-      name: "Marex Spectron Europe Limited - Otf",
-      acronym: "MSEL OTF",
-      micCode: "MSEL",
-      country: "Ireland",
-      currency: "Euro",
-      timeZone: "Europe/Dublin",
-      openTime: "08:00",
-      closeTime: "16:30"
-    },
-    {
-      name: "Borsa Italiana Equity Mtf",
-      acronym: "BITEQMTF",
-      micCode: "MTAH",
-      country: "Italy",
-      currency: "Euro",
-      timeZone: "Europe/Rome",
-      openTime: "09:00",
-      closeTime: "17:25"
-    },
-    {
-      name: "Clearcorp Dealing Systems India Limited - Astroid",
-      acronym: "ASTROID",
-      micCode: "ASTR",
-      country: "India",
-      currency: "Indian Rupee",
-      timeZone: "Asia/Kolkata",
-      openTime: "09:15",
-      closeTime: "15:30"
-    },
-    {
-      name: "Memx Llc Equities",
-      acronym: "MEMX",
-      micCode: "MEMX",
-      country: "United States",
-      currency: "United States Dollar",
-      timeZone: "America/New_York",
-      openTime: "09:30",
-      closeTime: "16:00"
-    },
-    {
-      name: "Natixis - Systematic Internaliser",
-      acronym: "NATX",
-      micCode: "NATX",
-      country: "France",
-      currency: "Euro",
-      timeZone: "Europe/Paris",
-      openTime: "09:00",
-      closeTime: "17:30"
-    },
-    {
-      name: "Currenex Ireland Mtf - Rfq",
-      acronym: "CNX MTF",
-      micCode: "ICXR",
-      country: "Ireland",
-      currency: "Euro",
-      timeZone: "Europe/Dublin",
-      openTime: "08:00",
-      closeTime: "16:30"
-    },
-    {
-      name: "Neo Exchange - Neo-l (market By Order)",
-      acronym: "NEO-L",
-      micCode: "NEOE",
-      country: "Canada",
-      currency: "Canadian Dollar",
-      timeZone: "America/Montreal",
-      openTime: "09:30",
-      closeTime: "16:00"
-    },
-    {
-      name: "Polish Trading Point",
-      acronym: "PTP",
-      micCode: "PTPG",
-      country: "Poland",
-      currency: "Polish Zloty",
-      timeZone: "Europe/Warsaw",
-      openTime: "09:00",
-      closeTime: "17:35"
-    },
-    {
-      name: "Pfts Stock Exchange",
-      acronym: "PFTS",
-      micCode: "PFTS",
-      country: "Ukraine",
-      currency: "Ukrainian Hryvnia",
-      timeZone: "Europe/Kiev",
-      openTime: "10:00",
-      closeTime: "17:30"
-    }
-  ];
-
